requirements.txt
unstructured
tiktoken
pinecone-client
pypdf
openai
langchain
pandas 
numpy  
python-dotenv